Safari-609.1.11.1/LayoutTests/security/isSecureContext-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ enableIsSecureContextAttribute=false ] -->
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
window.jsTestIsAsync = true;

description("Tests that global property isSecureContexts is undefined when the Setting isSecureContextAttributeEnabled is disabled.");

shouldBeUndefined("window.isSecureContext");
startWorker("resources/worker-isSecureContext-disabled.js");
</script>
</body>
</html>
