Safari-609.1.11.1/LayoutTests/security/isSecureContext-disabled-expected.txt

Tests that global property isSecureContexts is undefined when the Setting isSecureContextAttributeEnabled is disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.isSecureContext is undefined.
Starting worker: resources/worker-isSecureContext-disabled.js
PASS [Worker] self.isSecureContext is undefined.
PASS successfullyParsed is true

TEST COMPLETE
